fix: address review findings from phase 1
- AuditLog.id: Integer -> BigInteger in ORM
- allowed_hosts: secure default ["localhost", "127.0.0.1"]
- Migration 0002: CHECK constraints for users.role, filters.match_operator, filter_actions.action_type
- Migration 0003: singleton constraint on app_settings (id = 1)
- Move FetchLog from article.py to dedicated fetch_log.py
- CSRF: deferred to phase 2 (depends on auth cookie strategy)

+++ backend/app/models/settings.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Boolean, DateTime, Integer, SmallInteger, String, Text, ForeignKey, func
+from sqlalchemy import BigInteger, Boolean, DateTime, Integer, SmallInteger, String, Text, ForeignKey, func
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -29,7 +29,7 @@ class AppSettings(Base):
 class AuditLog(Base):
     __tablename__ = "audit_log"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     admin_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id", ondelete="RESTRICT"), nullable=False)
     action: Mapped[str] = mapped_column(String(50), nullable=False)
     target_type: Mapped[str | None] = mapped_column(String(30))
